Make NoSolutionError pickleable
Sometimes you have an unsolveable optimization problem running in a
subprocess. It's useful to allow NoSolutionError to be pickleable so
that the exception can propagate back to the main process.

Without this change, unpickling the exception fails. BaseException has
an implementation of __reduce__, which defines the state to pickle,
which remembers the args that its constructor received. Upon
deserialization, the constructor of the exception is called with the
arguments that the BaseException constructor received.

If the BaseException constructor does not receive NoSolutionError's
mandatory `problem` argument, we will try to call the NoSolutionError
constructor without it, which will fail.

An alternative is reimplementing __reduce__ on NoSolutionError.